+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Support setting database flags (new configuration parameter: `database_flags`).
+- Support setting the storage auto resize limit (new configuration parameter: `auto_resize_limit`).
 
 ## [4.3.0] - 2019-08-26